+++ CHANGES.rst
@@ -7,6 +7,8 @@ Products.MeetingCharleroi Changelog
 
 - Make `no_freeze` and `no_decide` WF adaptations available.
   [gbastien]
+- Adapted code for `ToolPloneMeeting.advisersConfig`, added upgrade step to `4201`.
+  [gbastien]
 
 4.2.0rc2 (2023-10-20)
 ---------------------